chore:更新参数检查逻辑以支持更大数值范围
在 `ArgumentChecker.cs` 中，将 `ThrowIfLEZero` 方法的参数类型从 `int` 更改为 `long`，以支持更大的数值范围，并保持原有逻辑不变。

在 `AutoMapperObject.cs` 中，将 `Map` 方法中的 `id` 参数检查逻辑更新为使用 `Checker.Argument.ThrowIfLEZero`，以确保 `id` 为有效的正数。

+++ src/Infrastructures/Core/Guard/ArgumentChecker.cs
@@ -20,7 +20,7 @@ public class ArgumentChecker
     /// <param name="argument">Argument.</param>
     /// <param name="argumentName">Argument name.</param>
     /// <param name="message">message.</param>
-    public void ThrowIfLEZero(int argument, string argumentName, string? message = null)
+    public void ThrowIfLEZero(long argument, string argumentName, string? message = null)
     {
         if (argument <= 0)
         {
